Refactor to traits.
